+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -2,14 +2,15 @@
 package com.unciv.build
 
 object BuildConfig {
-    const val kotlinVersion = "1.9.24"
+    const val kotlinVersion = "2.1.20"
     const val appName = "Unciv"
-    const val appCodeNumber = 1141
-    const val appVersion = "4.17.3"
+    const val appCodeNumber = 1144
+    const val appVersion = "4.17.6"
 
     const val gdxVersion = "1.13.1"
-    const val ktorVersion = "2.3.12"
+    const val ktorVersion = "2.3.13"
     const val coroutinesVersion = "1.8.1"
+    const val jnaVersion = "5.17.0"
 
     const val identifier = "com.unciv.app"
 }