+++ core/src/com/unciv/logic/civilization/managers/ImprovementFunctions.kt
@@ -4,7 +4,7 @@ import com.unciv.logic.civilization.Civilization
 import com.unciv.logic.map.tile.ImprovementBuildingProblem
 import com.unciv.logic.map.tile.Tile
 import com.unciv.models.ruleset.tile.TileImprovement
-import com.unciv.models.ruleset.unique.StateForConditionals
+import com.unciv.models.ruleset.unique.GameContext
 import com.unciv.models.ruleset.unique.UniqueType
 
 object ImprovementFunctions {
@@ -12,9 +12,9 @@ object ImprovementFunctions {
     /** Generates a sequence of reasons that prevent building given [improvement].
      *  If the sequence is empty, improvement can be built immediately.
      */
-    fun getImprovementBuildingProblems(improvement: TileImprovement, stateForConditionals: StateForConditionals, tile: Tile? = null): Sequence<ImprovementBuildingProblem> = sequence {
-        if (stateForConditionals.civInfo != null) {
-            val civInfo: Civilization = stateForConditionals.civInfo
+    fun getImprovementBuildingProblems(improvement: TileImprovement, gameContext: GameContext, tile: Tile? = null): Sequence<ImprovementBuildingProblem> = sequence {
+        if (gameContext.civInfo != null) {
+            val civInfo: Civilization = gameContext.civInfo
 
             if (improvement.uniqueTo != null && !civInfo.matchesFilter(improvement.uniqueTo!!))
                 yield(ImprovementBuildingProblem.WrongCiv)
@@ -22,24 +22,24 @@ object ImprovementFunctions {
                 yield(ImprovementBuildingProblem.Replaced)
             if (improvement.techRequired != null && !civInfo.tech.isResearched(improvement.techRequired!!))
                 yield(ImprovementBuildingProblem.MissingTech)
-            if (improvement.getMatchingUniques(UniqueType.Unbuildable, StateForConditionals.IgnoreConditionals)
+            if (improvement.getMatchingUniques(UniqueType.Unbuildable, GameContext.IgnoreConditionals)
                     .any { it.modifiers.isEmpty() })
                 yield(ImprovementBuildingProblem.Unbuildable)
-            else if (improvement.hasUnique(UniqueType.Unbuildable, stateForConditionals))
+            else if (improvement.hasUnique(UniqueType.Unbuildable, gameContext))
                 yield(ImprovementBuildingProblem.ConditionallyUnbuildable)
 
-            if (improvement.hasUnique(UniqueType.Unavailable, stateForConditionals))
+            if (improvement.hasUnique(UniqueType.Unavailable, gameContext))
                 yield(ImprovementBuildingProblem.ConditionallyUnbuildable)
 
-            if (improvement.getMatchingUniques(UniqueType.OnlyAvailable, StateForConditionals.IgnoreConditionals)
-                    .any { !it.conditionalsApply(stateForConditionals) })
+            if (improvement.getMatchingUniques(UniqueType.OnlyAvailable, GameContext.IgnoreConditionals)
+                    .any { !it.conditionalsApply(gameContext) })
                 yield(ImprovementBuildingProblem.UnmetConditional)
 
-            if (improvement.getMatchingUniques(UniqueType.ObsoleteWith, stateForConditionals)
+            if (improvement.getMatchingUniques(UniqueType.ObsoleteWith, gameContext)
                     .any { civInfo.tech.isResearched(it.params[0]) })
                 yield(ImprovementBuildingProblem.Obsolete)
 
-            if (improvement.getMatchingUniques(UniqueType.ConsumesResources, stateForConditionals)
+            if (improvement.getMatchingUniques(UniqueType.ConsumesResources, gameContext)
                     .any { civInfo.getResourceAmount(it.params[1]) < it.params[0].toInt() })
                 yield(ImprovementBuildingProblem.MissingResources)
 
@@ -49,8 +49,8 @@ object ImprovementFunctions {
                 yield(ImprovementBuildingProblem.MissingResources)
             
             if (tile != null) {
-                if (tile.getOwner() != civInfo && !improvement.hasUnique(UniqueType.CanBuildOutsideBorders, stateForConditionals)) {
-                    if (!improvement.hasUnique(UniqueType.CanBuildJustOutsideBorders, stateForConditionals))
+                if (tile.getOwner() != civInfo && !improvement.hasUnique(UniqueType.CanBuildOutsideBorders, gameContext)) {
+                    if (!improvement.hasUnique(UniqueType.CanBuildJustOutsideBorders, gameContext))
                         yield(ImprovementBuildingProblem.OutsideBorders)
                     else if (tile.neighbors.none { it.getOwner() == civInfo })
                         yield(ImprovementBuildingProblem.NotJustOutsideBorders)
@@ -60,7 +60,7 @@ object ImprovementFunctions {
                         rulesetImprovement.techRequired == null || civInfo.tech.isResearched(rulesetImprovement.techRequired!!)
                     }
         
-                if (!tile.improvementFunctions.canImprovementBeBuiltHere(improvement, tile.hasViewableResource(civInfo), knownFeatureRemovals, stateForConditionals))
+                if (!tile.improvementFunctions.canImprovementBeBuiltHere(improvement, tile.hasViewableResource(civInfo), knownFeatureRemovals, gameContext))
                 // There are way too many conditions in that functions, besides, they are not interesting
                 // at least for the current usecases. Improve if really needed.
                     yield(ImprovementBuildingProblem.Other)