+++ core/src/com/unciv/logic/civilization/managers/UnitManager.kt
@@ -13,6 +13,7 @@ import com.unciv.models.ruleset.unique.UniqueTarget
 import com.unciv.models.ruleset.unique.UniqueTriggerActivation
 import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.ruleset.unit.BaseUnit
+import yairm210.purity.annotations.Readonly
 
 class UnitManager(val civInfo: Civilization) {
 
@@ -126,9 +127,10 @@ class UnitManager(val civInfo: Civilization) {
         }
         return unit
     }
-    fun getCivUnitsSize(): Int = unitList.size
-    fun getCivUnits(): Sequence<MapUnit> = unitList.asSequence()
-    fun getCivGreatPeople(): Sequence<MapUnit> = getCivUnits().filter { mapUnit -> mapUnit.isGreatPerson() }
+
+    @Readonly fun getCivUnitsSize(): Int = unitList.size
+    @Readonly fun getCivUnits(): Sequence<MapUnit> = unitList.asSequence()
+    @Readonly fun getCivGreatPeople(): Sequence<MapUnit> = getCivUnits().filter { mapUnit -> mapUnit.isGreatPerson() }
 
     // Similar to getCivUnits(), but the returned list is rotated so that the
     // 'nextPotentiallyDueAt' unit is first here.