+++ core/src/com/unciv/models/translations/Translations.kt
@@ -14,6 +14,9 @@ import com.unciv.utils.Log
 import com.unciv.utils.debug
 import java.util.Locale
 import org.jetbrains.annotations.VisibleForTesting
+import yairm210.purity.annotations.LocalState
+import yairm210.purity.annotations.Pure
+import yairm210.purity.annotations.Readonly
 
 /**
  *  This collection holds all translations for the game.
@@ -472,10 +475,13 @@ private fun String.translateIndividualWord(language: String, hideIcons: Boolean,
  * For example, a string like 'The city of [New [York]]' will return ['New [York]'],
  * allowing us to have nested translations!
  */
+@Readonly
 fun String.getPlaceholderParameters(): List<String> {
     if (!this.contains('[')) return emptyList()
 
     val stringToParse = this.removeConditionals()
+    
+    @LocalState
     val parameters = ArrayList<String>()
     var depthOfBraces = 0
     var startOfCurrentParameter = -1
@@ -492,6 +498,7 @@ fun String.getPlaceholderParameters(): List<String> {
     return parameters
 }
 
+@Readonly
 fun String.getPlaceholderText(): String {
     var stringToReturn = this.removeConditionals()
     val placeholderParameters = stringToReturn.getPlaceholderParameters()
@@ -500,6 +507,7 @@ fun String.getPlaceholderText(): String {
     return stringToReturn
 }
 
+@Readonly
 fun String.equalsPlaceholderText(str: String): Boolean {
     if (isEmpty()) return str.isEmpty()
     if (str.isEmpty()) return false // Empty strings have no .first()
@@ -507,6 +515,7 @@ fun String.equalsPlaceholderText(str: String): Boolean {
     return this.getPlaceholderText() == str
 }
 
+@Pure
 fun String.hasPlaceholderParameters(): Boolean {
     if (!this.contains('[')) return false
     return squareBraceRegex.containsMatchIn(this.removeConditionals())
@@ -524,11 +533,13 @@ fun String.fillPlaceholders(vararg strings: String): String {
     return filledString
 }
 
+@Pure @Suppress("purity")
 fun String.getModifiers(): List<Unique> {
     if (!this.contains('<')) return emptyList()
     return pointyBraceRegex.findAll(this).map { Unique(it.groups[1]!!.value) }.toList()
 }
 
+@Pure
 fun String.removeConditionals(): String {
     if (!this.contains('<')) return this // no need to regex search
     return this