+++ core/src/com/unciv/ui/screens/worldscreen/unit/AutoPlay.kt
@@ -4,6 +4,7 @@ import com.unciv.models.metadata.GameSettings
 import com.unciv.ui.screens.worldscreen.WorldScreen
 import com.unciv.utils.Concurrency
 import kotlinx.coroutines.Job
+import yairm210.purity.annotations.Readonly
 
 class AutoPlay(private var autoPlaySettings: GameSettings.GameSettingsAutoPlay) {
     /**
@@ -61,6 +62,7 @@ class AutoPlay(private var autoPlaySettings: GameSettings.GameSettingsAutoPlay) 
         }
     }
 
+    @Readonly
     fun isAutoPlaying(): Boolean = turnsToAutoPlay > 0 || autoPlayTurnInProgress
 
     fun isAutoPlayingAndFullAutoPlayAI(): Boolean = isAutoPlaying() && autoPlaySettings.fullAutoPlayAI